Fix ambiguous conditional expression w/ Clang+Windows

+++ src/time_zone_format.cc
@@ -48,7 +48,8 @@ char* strptime(const char* s, const char* fmt, std::tm* tm) {
   std::istringstream input(s);
   input >> std::get_time(tm, fmt);
   if (input.fail()) return nullptr;
-  return const_cast<char*>(s) + (input.eof() ? strlen(s) : input.tellg());
+  return const_cast<char*>(s) +
+         (input.eof() ? strlen(s) : static_cast<std::size_t>(input.tellg()));
 }
 #endif
